[native_titlebar] add support for popouts
