+++ branches/tweak/native_titlebar/meta.js
@@ -1,2 +1,2 @@
-export const name = "Windows Native Titlebar";
+export const name = "Native Titlebar";
 export const description = "Replaces Discord's custom titlebar with Windows' native one";
