--- branches/tweak/windows_native_titlebar/preload.js
@@ -1,3 +1,0 @@
-const { webFrame } = require("electron");
-
-webFrame.insertCSS("[class*=titleBar]{display: none!important}");
